Added Reroute Capabilities
